requirements.txt
Flask==2.1.1
tensorflow==2.8.0
Pillow==9.1.0
gunicorn==20.1.0
